LegalAssistant/requirements.txt
# Frontend
flask
flask-cors
faiss-cpu
# Model
sentence-transformers
PyMuPDF
transformers
accelerate
torch
numpy
faiss-cpu
langchain
langchain-community
langchain-huggingface
langchain-ollama
ollama